# contracts/backend-openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "Reproit Cloud backend contract", "version": "0.1.0-experimental" },
  "paths": {
    "/auth/signup": { "post": { "operationId": "cloudSignup", "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Credentials" } } } }, "responses": { "201": { "description": "created", "content": { "application/json": { "schema": { "type": "object", "required": ["email"], "additionalProperties": false, "properties": { "email": { "type": "string", "format": "email" } } } } } } } } },
    "/account/projects": { "post": { "operationId": "cloudCreateProject", "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["name"], "additionalProperties": false, "properties": { "name": { "type": "string", "minLength": 1, "maxLength": 80 } } } } } }, "responses": { "201": { "description": "created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreatedProject" } } } } } } },
    "/v1/events": { "post": { "operationId": "cloudIngestEvents", "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["appId", "events"], "properties": { "appId": { "type": "string", "minLength": 1 }, "batchId": { "type": "string", "minLength": 1, "maxLength": 128 }, "events": { "type": "array", "maxItems": 5000, "items": {} }, "ctx": { "type": "object" } } } } } }, "responses": { "200": { "description": "accepted", "content": { "application/json": { "schema": { "oneOf": [{ "$ref": "#/components/schemas/IngestedBatch" }, { "$ref": "#/components/schemas/DedupedBatch" }] } } } } } } },
    "/v1/me": { "get": { "operationId": "cloudGetMe", "responses": { "200": { "description": "identity", "content": { "application/json": { "schema": { "type": "object", "required": ["orgId", "projects"], "additionalProperties": false, "properties": { "orgId": { "type": "integer" }, "projects": { "type": "integer", "minimum": 0 } } } } } } } } },
    "/v1/apps/{app}/buckets/{bucket}/replay-results": { "post": { "operationId": "cloudRecordReplay", "parameters": [{ "in": "path", "name": "app", "required": true, "schema": { "type": "string" } }, { "in": "path", "name": "bucket", "required": true, "schema": { "type": "string" } }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ReplayResultInput" } } } }, "responses": { "200": { "description": "recorded", "content": { "application/json": { "schema": { "type": "object", "required": ["ok", "id"], "additionalProperties": false, "properties": { "ok": { "const": true }, "id": { "type": "integer" } } } } } } } } }
  },
  "components": { "schemas": {
    "Credentials": { "type": "object", "required": ["email", "password"], "additionalProperties": false, "properties": { "email": { "type": "string", "format": "email", "maxLength": 254 }, "password": { "type": "string", "minLength": 8 } } },
    "CreatedProject": { "type": "object", "required": ["id", "name", "appId", "apiKey", "apiKeyPrefix", "publishableKey", "publishableKeyPrefix"], "additionalProperties": false, "properties": { "id": { "type": "integer" }, "name": { "type": "string" }, "appId": { "type": "string" }, "apiKey": { "type": "string" }, "apiKeyPrefix": { "type": "string" }, "publishableKey": { "type": "string" }, "publishableKeyPrefix": { "type": "string" } } },
    "IngestedBatch": { "type": "object", "required": ["ok", "ingested"], "properties": { "ok": { "const": true }, "ingested": { "type": "object", "required": ["edges", "errors"], "additionalProperties": false, "properties": { "edges": { "type": "integer", "minimum": 0 }, "errors": { "type": "integer", "minimum": 0 } } }, "droppedUntagged": { "type": "integer", "minimum": 1 } } },
    "DedupedBatch": { "type": "object", "required": ["ok", "deduped"], "additionalProperties": false, "properties": { "ok": { "const": true }, "deduped": { "const": true } } },
    "ReplayResultInput": { "type": "object", "required": ["status"], "properties": { "status": { "type": "string", "enum": ["reproduced", "clean", "data_dependent", "stale", "flaky"] }, "runs": { "type": "integer", "minimum": 0 }, "failures": { "type": "integer", "minimum": 0 }, "localReproId": { "type": "string" }, "fixedInBuild": { "type": "string" }, "runId": { "type": "integer" } } }
  } }
}